# medibuddy-backend/MediBuddy-Backend-main/openapi.yaml
openapi: 3.0.3
info:
  title: MediBuddy API
  description: |
    REST API for MediBuddy — a medication management, symptom logging, and
    consultation-summary PWA designed for low-resource settings.

    ## Base URL
    All endpoints are prefixed with `/api`.

    ## Authentication
    Most endpoints require a **Bearer JWT** in the `Authorization` header.
    Obtain tokens via the OTP → verify flow (see Auth endpoints).

    ## Response Envelope
    Every response follows:
    ```json
    { "success": true, "data": { ... }, "message": "optional", "serverTime": "ISO-8601" }
    ```
    Errors:
    ```json
    { "success": false, "error": "msg", "code": "CODE", "details": [], "serverTime": "ISO-8601" }
    ```

    ## Idempotency
    Mutating endpoints accept an `Idempotency-Key` header (UUID v4).
    The server caches responses for 24 h; replaying the same key returns the cached result.

    ## Rate Limits
    | Preset | Window | Max |
    |--------|--------|-----|
    | General | 15 min | 100 |
    | Auth | 15 min | 20 |
    | OTP | 15 min | 5 |
    | Sync | 1 min | 30 |
    | Write | 1 min | 60 |

    Rate headers: `X-RateLimit-Limit`, `X-RateLimit-Remaining`, `Retry-After`.

  version: 1.0.0
  contact:
    name: MediBuddy Team
  license:
    name: UNLICENSED
servers:
  - url: http://localhost:5000/api
    description: Local development
  - url: https://medibuddy-backend-w9hf.onrender.com/api
    description: Production (Render)
  - url: http://localhost:4010
...
paths:
...
  /profiles/{profileId}/reminders:
    parameters:
      - $ref: "#/components/parameters/profileId"
    get:
      operationId: listReminders
      tags: [Reminders]
      summary: List reminders (24 h window)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Array of reminder events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /profiles/{profileId}/reminders/{reminderId}/ack:
    parameters:
      - $ref: "#/components/parameters/profileId"
      - name: reminderId
        in: path
        required: true
        schema:
          type: string
    post:
      operationId: ackReminder
      tags: [Reminders]
      summary: Acknowledge a reminder (taken / snooze / skip)
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action:
                  type: string
                  enum: [taken, snooze, skip]
                deviceId:
                  type: string
                clientTs:
                  type: string
                  format: date-time
            example:
              action: "taken"
              deviceId: "browser-xyz"
              clientTs: "2026-02-10T08:01:00.000Z"
      responses:
        "200":
          description: Reminder acknowledged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"

  # ═══════ SYMPTOMS ═══════
  /profiles/{profileId}/symptoms:
    parameters:
      - $ref: "#/components/parameters/profileId"
    get:
      operationId: listSymptoms
      tags: [Symptoms]
      summary: List symptom logs with optional date range
      security:
        - bearerAuth: []
      parameters:
        - name: start
          in: query
          schema:
            type: string
            format: date-time
        - name: end
          in: query
          schema:
            type: string
            format: date-time
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated symptoms
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
    post:
      operationId: logSymptom
      tags: [Symptoms]
      summary: Log a symptom (supports idempotency)
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSymptom"
      responses:
        "201":
          description: Symptom logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"

  # ═══════ SUMMARIES ═══════
  /profiles/{profileId}/summaries:
    parameters:
      - $ref: "#/components/parameters/profileId"
    get:
      operationId: listSummaries
      tags: [Summaries]
      summary: List consultation summaries
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Array of summaries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
    post:
      operationId: generateSummary
      tags: [Summaries]
      summary: Generate a consultation summary PDF
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [startDate, endDate]
              properties:
                startDate:
                  type: string
                  format: date
                endDate:
                  type: string
                  format: date
            example:
              startDate: "2026-01-01"
              endDate: "2026-02-01"
      responses:
        "201":
          description: Summary generation started / completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /profiles/{profileId}/summaries/{summaryId}:
    parameters:
      - $ref: "#/components/parameters/profileId"
      - name: summaryId
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getSummary
      tags: [Summaries]
      summary: Get summary details and PDF URL
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Summary with signed PDF download URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
              example:
                success: true
                data:
                  _id: "sum001"
                  profileId: "aaa111"
                  status: "completed"
                  pdfUrl: "https://b2.example.com/signed-url..."
                  startDate: "2026-01-01"
                  endDate: "2026-02-01"
                  medicationCount: 3
                  symptomCount: 12
                serverTime: "2026-02-10T12:10:00.000Z"
  /profiles/{profileId}/summaries/{summaryId}/download:
    parameters:
      - $ref: "#/components/parameters/profileId"
      - name: summaryId
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: downloadSummary
      tags: [Summaries]
      summary: Download summary as PDF (inline)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: PDF binary
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "404":
          $ref: "#/components/responses/NotFound"

  # ═══════ CAREGIVERS ═══════
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Obtain via `POST /auth/verify`. Short-lived access token (15 min).
        Use `POST /auth/refresh` with the refresh token to get a new pair.
  parameters:
    profileId:
      name: profileId
      in: path
      required: true
      schema:
        type: string
      description: MongoDB ObjectId of the profile
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        format: uuid
      description: UUID v4 — server caches responses for 24 h
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            success: false
            error: "Resource not found"
            code: "NOT_FOUND"
            serverTime: "2026-02-10T12:00:00.000Z"
...
  schemas:
    SuccessEnvelope:
      type: object
      required: [success, serverTime]
      properties:
        success:
          type: boolean
          example: true
        data:
          description: Response payload (varies per endpoint)
        message:
          type: string
        serverTime:
          type: string
          format: date-time
    ErrorEnvelope:
      type: object
      required: [success, error, code, serverTime]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        code:
          type: string
        details:
          type: array
          items:
            type: object
            properties:
              path:
                type: string
              message:
                type: string
        serverTime:
          type: string
          format: date-time
...
    CreateSymptom:
      type: object
      required: [symptom, severity, clientTs]
      properties:
        symptom:
          type: string
          example: "headache"
        severity:
          type: integer
          minimum: 1
          maximum: 5
          example: 3
        duration:
          type: string
          example: "2 hours"
        notes:
          type: string
        voiceNoteUrl:
          type: string
        clientId:
          type: string
        clientTs:
          type: string
          format: date-time
...